requirements.txt

numpy==2.3.1
pillow==11.2.1
progress==1.6
